+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 3.1.0
+- Expose `uri` for file payloads and deprecated `filepath`.
+- Added convinience method `copyFileAndDeleteOriginal` for moving file using
+`uri`.
+
 ## 3.0.1
 - Fix issues when running with sound null safety.